refactor: Rename methods in CampeonatoRepository
This commit refactors the methods in the `CampeonatoRepository` class to follow consistent naming conventions. The `getCampeonato` method is renamed to `getCampeonatos` to reflect that it retrieves all championships. The `createcampeonato` method is renamed to `createCampeonato` to use camel case. The `updatecampeonato` method is removed as it is not used. These changes improve code readability and maintainability.