refactor: Add ModalidadesRepository
This commit adds the ModalidadesRepository class, which provides methods for retrieving, creating, updating, and deleting modalidades (types of competitions) in the system. It also includes methods for retrieving modalidades by ID, name, campeonato ID, and type. This repository will be used to interact with the modalidades table in the database.